+++ ms-catalogo/src/main/java/com/tpi/mscatalogo/dto/DepositoDTO.java
@@ -1,5 +1,7 @@
 package com.tpi.mscatalogo.dto;
 
+import java.math.BigDecimal;
+
 public class DepositoDTO {
 
     private Long id;
@@ -7,8 +9,8 @@ public class DepositoDTO {
     private String direccion;
     private String localidad;
     private String provincia;
-    private Double latitud;
-    private Double longitud;
+    private BigDecimal latitud;
+    private BigDecimal longitud;
     private Integer capacidadMaxima;
     private Double costoEstadiaDiaria;
 
@@ -27,11 +29,11 @@ public class DepositoDTO {
     public String getProvincia() { return provincia; }
     public void setProvincia(String provincia) { this.provincia = provincia; }
 
-    public Double getLatitud() { return latitud; }
-    public void setLatitud(Double latitud) { this.latitud = latitud; }
+    public BigDecimal getLatitud() { return latitud; }
+    public void setLatitud(BigDecimal latitud) { this.latitud = latitud; }
 
-    public Double getLongitud() { return longitud; }
-    public void setLongitud(Double longitud) { this.longitud = longitud; }
+    public BigDecimal getLongitud() { return longitud; }
+    public void setLongitud(BigDecimal longitud) { this.longitud = longitud; }
 
     public Integer getCapacidadMaxima() { return capacidadMaxima; }
     public void setCapacidadMaxima(Integer capacidadMaxima) { this.capacidadMaxima = capacidadMaxima; }
